+++ tests/tests.rs
@@ -28,6 +28,7 @@ mod test_filter;
 mod test_fixlengths;
 mod test_fmt;
 mod test_frequency;
+mod test_from;
 mod test_groupby;
 mod test_headers;
 mod test_implode;